Tooling to support creation of 4D datasets
If we want to test our code against datasets where Z is not provided, we need to be able to build synthetic datasets where Z is not provided.  This is the tooling to make such datasets (not ImageStacks!).  This means the TileSet is fundamentally 4D, with no zplanes, and no z coordinates.